Moved photo upload components to profile folder.
