+++ src/scss/upload-photo.scss
@@ -1,5 +1,5 @@
 .Dropzone {
-  height: 216px;
+  height: 150px;
 	width: 216px;
 	position: absolute;
 	top: 16px;
